+++ packages/seroval/src/core/base-parser/stream.ts
