+++ packages/seroval/src/core/base-parser/sync.ts
